requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.2.5
django-crispy-forms==2.0
django-dotenv==1.4.2
django-utils-six==2.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
install==1.3.5
joblib==1.3.2
kombu==5.3.2
multidict==6.0.4
nltk==3.8.1
numpy==1.26.0
outcome==1.2.0
packaging==23.2
pandas==2.1.1
prompt-toolkit==3.0.39
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
textblob==0.17.1
to==0.3
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
webdriver-manager==4.0.1
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
